[DF-6261] 0.0005 replaced with 0.0005f
Former-commit-id: 609d25880a913d19d8dc4d0b2917e283f91a8c40

+++ Tools/QuickEd/Classes/EditorSystems/TransformSystem.cpp
@@ -112,7 +112,7 @@ struct MagnetLine
     Vector2::eAxis axis;
 };
 
-const float32 TRANSFORM_EPSILON = static_cast<float32>(0.0005);
+const float32 TRANSFORM_EPSILON = 0.0005f;
 } //namespace
 
 TransformSystem::TransformSystem(EditorSystemsManager* parent)
